requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
yfinance>=0.2.30
scipy>=1.12.0
numpy>=1.26.0
plotly>=5.18.0
fastapi>=0.115.0
uvicorn>=0.32.0
